# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: nexus 统一 AI 网关 —— API 契约
  # ⚠️ 本字段是 **OpenAPI 契约文档自身的版本**（OpenAPI 规范要求 info.version 必填），
  #    不是后端产物版本 —— 别拿它去断言 GET /api/health 的 data.version。
  #    后端版本唯一真源 = backend/pom.xml 的 <revision>（运行时见 paths./api/health 里 data.version 的说明）；
  #    0.2.0 是示例值，非真源；实际以 /api/health 的实际返回为准。
  version: 0.2.0
  description: |
    演示项目后端接口契约。阶段0：健康检查；阶段1：多租户 + 认证（登录/登出/取当前用户）；
    阶段2：统一 AI 网关（SSE 流式对话，见 `paths./api/chat/stream`）。

    ## 统一响应体约定
    所有接口（含错误）一律返回 `Result<T>`：

    ```json
    { "code": 0, "msg": "success", "data": { /* 各接口自定义 */ } }
    ```

    - `code === 0`：成功，`data` 为业务数据（`Result.data` 的类型见各接口 schema）；
    - `code !== 0`：失败，`msg` 是可直接展示给前端的文案，`data` 通常为 `null`
      （例外：健康检查 503 时 `data` 仍返回完整报告，便于前端定位故障依赖）。

    ## 错误码（阶段0 + 阶段1 已落地）
    | code | 常量 | HTTP | 含义 |
    | --- | --- | --- | --- |
    | 0 | SUCCESS | 200 | 成功 |
    | 10000 | BUSINESS_ERROR | 200 | 通用业务失败（`BusinessException` 默认码） |
    | 10100 | LOGIN_FAILED | 200 | 用户名或密码错误（**刻意不区分**二者，防账号枚举） |
    | 10101 | ACCOUNT_DISABLED | 200 | 账号已停用（口令正确后才可能返回） |
    | 10102 | TENANT_DISABLED | 200 | 租户已停用 |
    | 20000 | SERVICE_UNAVAILABLE | 503 | 依赖服务不可用（健康检查探活失败） |
    | 40100 | UNAUTHENTICATED | 401 | 未携带 token |
    | 40101 | TOKEN_INVALID | 401 | 登录状态无效（伪造 / 签名错 / Redis 白名单已无该 jti） |
    | 40102 | TOKEN_EXPIRED | 401 | 登录已过期（重新登录即可） |
    | 40400 | NOT_FOUND | 404 | 请求的路径/资源不存在 |
    | 50000 | SYSTEM_ERROR | 500 | 系统异常（详情仅进服务端日志，不返回给前端） |

...

...

    | code | 常量 | HTTP | 含义 |
...

...

...

...
servers:
...
security:
  # 全局默认：需要 Bearer token（阶段1 起"默认拒绝"）
  # 白名单路径（/api/health、/api/auth/login、/actuator/**）在各自的 operation 上覆盖为 security: []
  - bearerAuth: []
...
paths:
...
  /api/chat/stream:
    post:
      tags:
        - Chat
      operationId: streamChat
      summary: 流式对话（SSE）
      description: |
        把会话历史全量上送，后端按 `modelType` 路由到本地 Ollama 或云端 DeepSeek，
        并以 **SSE 事件流**逐片返回生成的文本（打字机效果）。**无状态**：不落库、不建会话表，
        多轮上下文由前端持有 `messages` 数组每次全量上送（刷新即丢）。

        ⚠️ **本接口的响应体不是一次性 JSON**，因此 OpenAPI 只能把它声明为
        `text/event-stream` + `type: string` —— **这是 OpenAPI 3.0.3 的规范限制，不是遗漏**。
        帧的真实结构（四种事件、每条时序规则、失败形态总表）以
        **`docs/api/README.md` §6** 为准，实现与解析约定在设计文档 `docs/design/02-统一AI网关.md` §3。

        ## 事件帧（四种，顺序固定）

        ```
        event: meta
        data: {"code":0,"msg":"success","data":{"modelType":"OLLAMA","model":"qwen2.5:7b","servedBy":"user-selected"}}

        event: delta
        data: {"code":0,"msg":"success","data":{"content":"你"}}

        event: done
        data: {"code":0,"msg":"success","data":{"finishReason":"stop","deltaCount":2,"durationMs":842}}
        ```

        | 事件 | 时机 | `data` 载荷 |
        | --- | --- | --- |
        | `meta` | 第一帧，仅有且必有一帧（在首个 `delta` 之前，只在首次取到上游分片后发出） | `modelType` / `model` / `servedBy` |
        | `delta` | 0~N 帧 | `content`（**增量**片段，不是累积全文 → 前端做追加） |
        | `done` | 正常结束（与 `error` 二选一，有且仅有一个） | `finishReason` / `deltaCount` / `durationMs` |
        | `error` | 异常结束 | `{"deltaCount": N}`（**对象**，不是 `null`） |

        > 上表四种帧的 `data`，在 `components/schemas` 里**各有一个**对应 schema：
        > `ChatStreamMeta` / `ChatStreamDelta` / `ChatStreamDone` / `ChatStreamError`。
        > （OpenAPI 3.0.3 无法把 schema 挂到 SSE 事件上，所以这一行就是"帧表 ↔ schema"的对应关系；
        > 少一个 schema 就意味着按本文件生成类型的人会漏掉那种帧。）

        **每帧的 `data` 仍是统一响应体 `Result<T>`**（`code` / `msg` / `data` 三字段），
        前端可复用与普通接口同一套解包与成功判定（设计决策 D4）。

        ## 三条解析规则（前后端都要遵守）

        1. `meta` 的「开流前」指**对客户端开流前**、非「调用上游前」：在首个上游分片之后、
           第一个 `delta` 之前发出，且只发一次；
        2. 终止帧唯一：`done` **xor** `error`；
        3. 取 `event:` / `data:` 后的值时**先剥掉至多一个前导空格**再 trim ——
           按示例字面 `slice(6)` 会因一个空格匹配不到事件名（页面空白而 curl 正常）；
           行终止符 `\r\n` / `\r` / `\n` 都合法，入缓冲区时先归一。

        ## 失败形态（**全异步口径**）

        `SseEmitter` 一返回，`200 + text/event-stream` 的响应头**已经发出去了**，此后状态码无法再改：

        | 失败点 | 载体 | HTTP | code |
        | --- | --- | --- | --- |
        | 无 token / 过期 / 伪造 | `Result` | 401 | 40100 / 40102 / 40101 |
        | 请求体校验失败 / JSON 畸形 | `Result` | 200 | 40001 |
        | `modelType` 取值未知 | `Result` | 200 | 10200 |
        | **模型线程池已满** | `Result` | **503** | 20100 |
        | 上游不可达 / 超时 / 中途报错 | `event: error` 帧 | *(已是 200)* | 20100 |
        | 未预期的内部异常 | `event: error` 帧 | *(已是 200)* | 50000 |
        | 客户端主动断开 | 无（连接已没了） | — | 服务端据此取消上游调用 |

        前端须同时处理 `20100` 的两种载体：**先看 `Content-Type`**
        （用 `includes('application/json')` 判断，Spring 会给 `text/event-stream` 带 `;charset=UTF-8`，全等比较会误判），
        再按帧解析。流的出口有三个：收到 `done`、收到 `error`、**或底层流 EOF**
        （后端崩溃/nginx 断流时不会有终止帧，只认前两个会让页面永久停在"生成中"）。
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              singleTurn:
                summary: 单轮提问，指定本地模型
                value:
                  messages:
                    - role: user
                      content: 用三句话介绍杭州
                  modelType: OLLAMA
              multiTurn:
                summary: 多轮（历史全量上送）；`modelType` 缺省 = 由后端决定
                value:
                  messages:
                    - role: user
                      content: 你好
                    - role: assistant
                      content: 你好！有什么可以帮你的？
                    - role: user
                      content: 介绍一下你自己
      responses:
        '200':
          description: |
            成功开流：`Content-Type: text/event-stream`（Spring 会带上 `;charset=UTF-8`），
            响应体是 §description 所述的事件流。
            ⚠️ **也为 `code=40001 / 10200` 的业务失败所用** —— 那两种情况下响应体是普通的
            `application/json` + `Result`（`code` 非 0），而不是事件流。
          content:
            text/event-stream:
              # 事件流在 OpenAPI 3.0.3 里没有原生表达方式：只能声明为字符串，
              # 帧的真实结构写在 description 与 docs/api/README.md §6.2。
              schema:
                type: string
                description: |
                  SSE 事件流文本。每帧形如 `event: <名>\ndata: <Result JSON>\n\n`，
                  事件名取值 `meta` / `delta` / `done` / `error`。
                  **请以 `docs/api/README.md` §6.2 为唯一事实源**，不要从本字段的类型反推结构。
              example: |
                event: meta
                data: {"code":0,"msg":"success","data":{"modelType":"OLLAMA","model":"qwen2.5:7b","servedBy":"user-selected"}}

                event: delta
                data: {"code":0,"msg":"success","data":{"content":"你"}}

                event: done
                data: {"code":0,"msg":"success","data":{"finishReason":"stop","deltaCount":1,"durationMs":842}}

            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              examples:
                paramInvalid:
                  summary: 请求体校验失败（HTTP 200，开流前）
                  value:
                    code: 40001
                    msg: 请求参数不合法
                    data: null
                modelUnsupported:
                  summary: modelType 未知取值（HTTP 200，开流前）
                  value:
                    code: 10200
                    msg: 不支持的模型类型
                    data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: |
            模型线程池已满（并发流上限），**开流前**即可判定，故能给出真正的 503。
            上游模型不可达**不会**走这里 —— 那时响应头已发，只能写 `event: error` 帧。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              examples:
                executorSaturated:
                  value:
                    code: 20100
                    msg: 模型服务暂时不可用，请稍后重试
                    data: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        登录接口返回的 `data.token`。用法：`Authorization: Bearer <token>`。
        服务端每次请求校验 ①JWT 签名与有效期 ②Redis 白名单中该 `jti` 是否仍在
        （登出/被清即为失效）。失败返回 HTTP 401 + `40100/40101/40102`。
  parameters:
    AuthorizationHeader:
      name: Authorization
      in: header
      required: true
      description: 形如 `Bearer <token>`（注意 Bearer 后有**一个空格**）
      schema:
        type: string
      example: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIn0.abc123
  responses:
    Unauthorized:
      description: |
        登录态缺失或失效。**响应体仍是统一 `Result`**（HTTP 401 是给前端 401 分支用的传输层信号）：

        | code | msg | 触发场景 |
        | --- | --- | --- |
        | 40100 | 未登录，请先登录 | 没带 `Authorization: Bearer` 头 |
        | 40101 | 登录状态无效，请重新登录 | token 签名/格式非法，或 Redis 白名单里已无该 jti（已登出 / Redis 被清 / Redis 不可用） |
        | 40102 | 登录已过期，请重新登录 | token 的 `exp` 已过 |

        前端处置：清 store 中的 token → 跳 `/login?redirect=<当前路径>`（并发多请求同时 401 时只跳一次）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          examples:
            unauthenticated:
              summary: 未携带 token（40100）
              value:
                code: 40100
                msg: 未登录，请先登录
                data: null
            tokenInvalid:
              summary: token 非法或已登出（40101）
              value:
                code: 40101
                msg: 登录状态无效，请重新登录
                data: null
            tokenExpired:
              summary: token 已过期（40102）
              value:
                code: 40102
                msg: 登录已过期，请重新登录
                data: null
  schemas:
...
    Result:
      type: object
      description: 统一响应体外壳，所有接口共用；`data` 的具体结构由各接口定义
      required:
        - code
        - msg
      properties:
        code:
          type: integer
          format: int32
          description: 响应码：0 = 成功，非 0 = 失败（错误码表见 info.description）
          example: 0
        msg:
          type: string
          description: 提示信息：成功固定为 `success`；失败为可直接展示给前端的文案
          example: success
        data:
          description: 业务数据；失败且无数据时为 `null`
          nullable: true

    # ── GET /api/health 的 data 定义（前端 HealthData）──
...
    ChatRequest:
      type: object
      description: |
        对话请求体（契约：`docs/api/README.md` §6.1）。**无状态**：多轮上下文由前端
        每次全量上送 `messages`，后端不落库（刷新即丢，前端应给用户提示）。
      required:
        - messages
      properties:
        messages:
          type: array
          minItems: 1
          description: |
            会话历史全量上送。至少 1 条，**最后一条必须是 `role=user`**。
            **允许连续同角色**：失败/停止后前端回滚该轮 assistant 消息后，数组会以 `user` 结尾，
            下一次提问再 push 一条 `user` 即成连续两条 —— 契约允许，无需特殊处理。
          items:
            $ref: '#/components/schemas/ChatMessage'
        modelType:
          type: string
          nullable: true
          description: |
            模型类型。**`null` 或缺省 = 由后端决定**（本轮解析为后端默认模型，`meta.servedBy` 会是 `default`）。
            未知取值 → `10200`（HTTP 200，开流前）。
            后端 DTO 刻意用 `String` 承接再由 `ModelType.parse()` 转枚举：若直接声明为枚举，
            Jackson 遇未知值会抛 `HttpMessageNotReadableException` → 500 + 50000，而不是契约要的 `10200`。
          enum:
            - OLLAMA
            - DEEPSEEK
          example: OLLAMA
    ChatMessage:
      type: object
      description: 一条会话消息（契约：`docs/api/README.md` §6.1）
      required:
        - role
        - content
      properties:
        role:
          type: string
          description: 消息角色。本阶段**不支持** `system`（提示词模板管理不在阶段2 范围内）
          enum:
            - user
            - assistant
        content:
          type: string
          description: |
            消息正文。非空、去空白后非空；**单条上限 8KB、所有消息合计 64KB**（按 UTF-8 字节数计，
            超出 → `40001`）。上限防的是误贴大段文本把上游上下文撑爆。
          example: 用三句话介绍杭州

    # ── 事件帧的 data 载荷（**四种帧各一个 schema**，与 README.md §6.2 的帧表逐项对应）──
    #    对应后端 dto/ChatStreamMeta / ChatStreamDelta / ChatStreamDone / ChatStreamError
    #    注意：这几个 schema 只描述帧内 data 的结构，完整帧形如 `event: X\ndata: Result<X>\n\n`
...